+++ src/workers/webdav/modules/files/test/infrastructure/content/FSContentsCacheRepository.test.ts
@@ -2,7 +2,7 @@ import glob from 'tiny-glob';
 import fs from 'fs/promises';
 import { createReadStream } from 'fs';
 import { Readable } from 'stream';
-import { FSContentsCacheRepository } from '../../../infrastructure/storage/FSContentsCacheRepository';
+import { FSContentsCacheRepository } from '../../../infrastructure/content/FSContentsCacheRepository';
 jest.mock('tiny-glob', () => jest.fn());
 jest.mock('fs/promises');
 jest.mock('fs');
